automatically appending dir separator does not work on Windows
A check for  path[0] != LT_DIR_SEPARATOR  is ALWAYS true on Windows,
appending LT_DIR_SEPARATOR_S to string to concatenate with path then
results in a final path of "\D:\..." that is invalid.

This change disables that automatism for Windows completely, if some
consumer really relied on it it would need a different algorithm to be
implemented.