do not specify a document encoding to avoid garbage conversion
libxml2 has a problem when a document encoding is specified, even though
documentation says that all internal representation would be in UTF-8
and language-subtag-registry data already is passed in UTF-8, when
specifying the document encoding to be UTF-8 the output to
language-subtag-registry.xml is conversion garbage for non-ASCII
characters, '?' on Linux UTF-8 locale, ISO-8859-1 on Windows. If no
document encoding is specified at all no conversion happens and
non-ASCII values are correctly written as entities.

This apparently is only a workaround, maybe proper libxml2 mechanisms
exist to solve the problem. However, without this the data is unusable
as an xml parser bails out when reading the ISO-8859-1 data on Windows.